fix: melhor organização das pastas e classes
