+++ src/maps/Graphics.java
@@ -1,0 +1,5 @@
+package maps;
+
+public class Graphics {
+
+}
